Progress on Issue#1490 "UmpleOnline module to allow running experiments"

This adds a major new feature that allows requestors to request modeling tasks of end-users and to have those results submitted by responders. This can be used by people teaching a course (for course assignments) or by experimenters.

+++ umpleonline/bookmark.php
@@ -13,8 +13,39 @@ $tempModelId = $_REQUEST["model"];
 // The following creates a random numbered directory in ump
 // the result is ump/{dir}/model.ump
 date_default_timezone_set('UTC');
-$savedModelData = dataStore()->createData(date("ymd"));
-$tempModelData = dataStore()->openData($tempModelId);
+
+if (isset($_REQUEST["taskname"])) {
+  $taskExist = false;
+  $savedModelData = dataStore()->createData("task-" . $_REQUEST["taskname"] . "-" . date("ymd"));
+  foreach (new DirectoryIterator("ump/tasks") as $file) 
+  {
+    if ($file->isDot()) continue;
+
+    if ($file->isDir() && substr($file->getFilename(), 0, 8) == "taskroot") 
+    {
+      $taskName = explode("-", $file->getFilename())[1];
+      if ($taskName == $_REQUEST["taskname"])
+      {
+        $tempModelId = $file->getFilename();
+        $taskExist = true;
+        break;
+      }
+    }
+  }
+
+  if ($taskExist)
+  {
+    $tempModelData = dataStore()->openData("tasks/" . $tempModelId);
+  }
+  else 
+  {
+    echo "Task with name '" . $_REQUEST["taskname"] . "' does not exist"; 
+    exit();
+  }
+} else {
+  $savedModelData = dataStore()->createData(date("ymd"));
+  $tempModelData = dataStore()->openData($tempModelId);
+}
 
 $saveModelId = $savedModelData->getName();
 
@@ -47,7 +78,19 @@ if (!$tempModelData->hasData('model.ump.erroroutput'))
 $savedModelData->cloneFrom($tempModelData);
 
 // Empty anything else in directory and remove it
-$tempModelData->delete();
+if (!isset($_REQUEST["taskname"])) 
+{
+  $tempModelData->delete();
+}
 
-header("Location: umple.php?model={$saveModelId}");
-
+if (isset($_REQUEST["loadTaskWithURL"]))
+{
+  header("Location: umple.php?model={$saveModelId}");
+}
+else if (isset($_REQUEST["taskname"])) {  
+  //header("Location: umple.php?model={$saveModelId}");
+  echo "$saveModelId";
+} 
+else {
+  header("Location: umple.php?model={$saveModelId}");
+}
